MethodErrorHints.jl

MethodErrorHints — Module

MethodErrorHints.jl

A Julia package providing a macro to register custom hint messages for MethodErrors.

This package exports a single macro, @method_error_hint, which allows you to register custom hint messages to be displayed when a MethodError is thrown for a function call.

MethodErrorHints.@method_error_hint — Macro

@method_error_hint sig msg [key=val...]

Register a hint message msg to be printed when a MethodError is thrown for a function call matching the signature sig.

The signature sig is specified in a similar way to a typical method definition in Julia. See below for examples.

Additional key-value pairs key=val are passed to Base.printstyled when printing the hint message, allowing customization of the appearance of the message. Please see the documentation of Base.printstyled for the available options.

Usage

This function uses Base.Experimental.register_error_hint, and thus all the suggestions in its docstring are also applicable here.

For example, this macro should be called within the __init__ function of a module, to ensure that the hint is registered when the module is loaded. Furthermore, since this is an experimental Julia feature, if you wish to guard against breakage you should gate calls to this macro behind if isdefined(Base.Experimental, :register_error_hint).

if isdefined(Base.Experimental, :register_error_hint)
    function __init__()
        @method_error_hint f1(x::Int, y; z::String) "My error hint" color=:red
    end
end

Examples

Here, ✔ indicates an invocation that will trigger the hint, and ✘ indicates one that will not. For the most part, the rules exactly mimic method dispatch in Julia, with some small exceptions noted in the next section.

In this first example, f1 must be called with two positional arguments and a keyword argument z that have precisely the types specified, plus a keyword argument z. Names of positional arguments are optional as long as a type is specified, but names of keyword arguments are mandatory. If a positional argument is untyped like y here, it is treated as Any.

function f1 end
@method_error_hint f1(x::Int, y; z::String) "My error hint" color=:red
# f1(3,   :a ; z="hello"     ) ✔
# f1(3,   2  ; z="hello"     ) ✔
# f1(3,   :a                 ) ✘ (no `z` keyword argument)
# f1(3.0, :a ; z="hello"     ) ✘ (type of `x` is wrong)
# f1(3,   :a ; z=1           ) ✘ (type of `z` is wrong)
# f1(3       ; z="hello"     ) ✘ (missing positional argument `y`)
# f1(3,   :a ; z="hello", w=1) ✘ (extra keyword argument)

Here, f2 must take a single positional argument of any type, and optionally a keyword argument z. If z is provided, it must be of type Int. The macro does not actually use the default value specified for the keyword argument; you can use undef to avoid specifying one.

function f2 end
@method_error_hint f2(x; z::Int=3) "Another error hint" bold=true
# f2(3                )  ✔
# f2(3     ; z=2      )  ✔
# f2(3     ; z="hello")  ✘ (type of `z` is wrong)
# f2(3, :a            )  ✘ (extra positional argument)
# f2(3, :a ; z=2      )  ✘ (extra positional argument)

Variable-length positional and keyword arguments are supported.

function f3 end
@method_error_hint f3(args...; kwargs...) "Another hint" italic=true
# Any invocation of `f3` will trigger this.

Known limitations

Vararg annotations on splatted positional arguments are not supported. This will probably be fixed in a future version.

Type parameters and where-clauses are not supported, for example f(x::T, y::T) where {T}. If you need this functionality, you should consider crafting your own error hint (for this particular signature, following the example in the docstring of Base.Experimental.register_error_hint, you could check that length(argtypes) == 2 && argtypes[1] == argtypes[2]). Support for type parameters will probably never be implemented in this package.

Extended help

Behind the scenes

This macro's only real job is to parse the method signature, and then generate a call to Base.Experimental.register_error_hint that checks that the method the user tried to invoke matches the signature. This is best illustrated by an example. This:

function f1 end
@method_error_hint f1(x::Int, y; z::String) "My error hint" color=:red

expands into something like this:

Base.Experimental.register_error_hint(MethodError) do io, exc, argtypes, kwargs
    # This Dict is parsed from the method signature passed to the macro, i.e., this is the
    # 'target' signature we are trying to match.
    target_kwargtypes = Dict{Symbol,Any}(:z => String)
    # This Vector is also parsed from the method signature, and indicates which keyword
    # arguments are mandatory (i.e., have no default value in the method signature).
    mandatory_kwargs = [:z]
    # This one indicates whether the method signature has `kwargs...`.
    has_varkwargs = false
    # Extract the actual symbols of the keyword arguments passed in the invocation
    kwarg_symbols = __kwargs_symbols(kwargs)
    if (
        # Check that the function name is correct
        exc.f === f1
        # Check that number of positional arguments is correct
        && length(argtypes) == 2
        # Check that types of positional arguments are correct
        && argtypes[1] <: Int
        && argtypes[2] <: Any
        # Check that any mandatory keyword arguments are present (i.e., if there is no
        # default value in the method signature, it must be present in the invocation)
        && haskey(kwarg_symbols, :z)
        # Check that no unexpected keyword arguments are present (i.e., any keyword argument
        # in the invocation must either be specified in the method signature and obey the
        # type specified in the signature, or the method signature must have `kwargs...`).
        && all(collect(kwargs)) do (sym, typ)
            if haskey(target_kwargtypes, sym)
                # On Julia <= 1.10 we get values instead of types so we have to use `isa`
                # instead of `<:`
                typ <: target_kwargtypes[sym]
            else
                has_varkwargs
            end
        end
    )
        printstyled(io, "My error hint"; :color=:red)
    end
end

If you would like to see this in more detail, please refer to the source code of this package.